Fix wrong figure being returned by findMouseEventTargetAt.
findMouseEventTargetInDescendantsAt used the singleton PRIVATE_POINT in
an unsafe way: It used it, even after it had been modified by a
recursive call to findMouseEventTargetAt (and thus
findMouseVentTargetInDescendantsAt).

This caused the problem that sometimes, the wrong figure would be
returned leading to mouse events being delivered to the wrong element.

For example, this bug is described in
https://github.com/ControlSystemStudio/cs-studio/issues/1818.